feat: refactor imports to use core.constants and add water levels command
