+++ db/aquifer_system.py
@@ -6,15 +6,14 @@ This is a master reference table for aquifer systems and hydrogeologic units.
 
 from typing import List, TYPE_CHECKING
 
-from sqlalchemy import Text, Index
-from sqlalchemy.orm import relationship, Mapped, mapped_column
-from sqlalchemy.ext.associationproxy import association_proxy, AssociationProxy
 from geoalchemy2 import Geometry
+from sqlalchemy import Text, Index
+from sqlalchemy.ext.associationproxy import association_proxy, AssociationProxy
+from sqlalchemy.orm import relationship, Mapped, mapped_column
 
+from core.constants import SRID_WGS84
 from db.base import Base, AutoBaseMixin, ReleaseMixin
 from db.lexicon import lexicon_term
-
-from constants import SRID_WGS84
 
 if TYPE_CHECKING:
     from db.thing import WellScreen, ThingAquiferAssociation, Thing
