+++ db/location.py
@@ -30,7 +30,7 @@ from sqlalchemy import (
 from sqlalchemy.ext.associationproxy import association_proxy, AssociationProxy
 from sqlalchemy.orm import relationship, Mapped, mapped_column
 
-from constants import SRID_WGS84
+from core.constants import SRID_WGS84
 from db.base import Base, AutoBaseMixin, ReleaseMixin
 from db.data_provenance import DataProvenanceMixin
 from db.notes import NotesMixin
