+++ pyproject.toml
@@ -100,6 +100,13 @@ dependencies = [
   "yarl==1.20.1",
 ]
 
+[tool.uv]
+package = true
+
+[tool.setuptools]
+packages = ["alembic", "cli", "core", "db", "schemas", "services"]
+modules = ["constants"]
+
 [project.scripts]
 oco = "cli.cli:cli"
