+++ schemas/location.py
@@ -14,19 +14,19 @@
 # limitations under the License.
 # ===============================================================================
 from datetime import date
+from typing import Any
 from typing import List
 
 from geoalchemy2 import WKBElement
 from geoalchemy2.shape import to_shape
 from pydantic import BaseModel, model_validator, field_validator, Field, ConfigDict
-from typing import Any
 
-from constants import SRID_WGS84, SRID_UTM_ZONE_13N
+from core.constants import SRID_WGS84, SRID_UTM_ZONE_13N
 from core.enums import ElevationMethod, CoordinateMethod
 from schemas import BaseCreateModel, BaseUpdateModel, BaseResponseModel
 from schemas.notes import NoteResponse, CreateNote, UpdateNote
-from services.validation.geospatial import validate_wkt_geometry
 from services.util import convert_m_to_ft, transform_srid
+from services.validation.geospatial import validate_wkt_geometry
 
 
 # -------- VALIDATE --------
